refactor(rr): place all RR-related code+tests in own rr subpackage
* renamed most "bui" dirs/package names to "rr.meteo"
* fnm/ toplevels files moved to rr/ (no separate subdir for fnm, open for discussion)
