--- docs/reference/bui.md
@@ -1,12 +1,0 @@
-# BUI file
-The "bui" file contains the precipitation input data for [Rainfall Runoff](glossary.md#rainfall-runoff).
-It is represented by the classes below.
-
-## Model
-::: hydrolib.core.io.bui.models
-
-## Parser
-::: hydrolib.core.io.bui.parser
-
-## Serializer
-::: hydrolib.core.io.bui.serializer
