+++ docs/reference/rr/meteo.md
@@ -1,0 +1,13 @@
+# The Rainfall Runoff meteo layer
+The meteo layer currently contains only support for the BUI file.
+The "bui" file contains the precipitation input data for [Rainfall Runoff](glossary.md#rainfall-runoff).
+It is represented by the classes below.
+
+## Model
+::: hydrolib.core.io.rr.meteo.models
+
+## Parser
+::: hydrolib.core.io.rr.meteo.parser
+
+## Serializer
+::: hydrolib.core.io.rr.meteo.serializer
