+++ hydrolib/core/io/rr/meteo/serializer.py
