+++ tests/test_model.py
@@ -19,11 +19,11 @@ from hydrolib.core.io.dimr.models import (
     StartGroup,
 )
 from hydrolib.core.io.ext.models import Boundary, ExtModel
-from hydrolib.core.io.fnm.models import RainfallRunoffModel
 from hydrolib.core.io.mdu.models import FMModel
+from hydrolib.core.io.rr.models import RainfallRunoffModel
 from hydrolib.core.io.xyz.models import XYZModel
 
-from .io.test_bui import BuiTestData
+from .io.rr.meteo.test_bui import BuiTestData
 from .utils import (
     assert_files_equal,
     invalid_test_data_dir,
